+++ .github/workflows/test-bot-contributor-lifecycle.yml
@@ -1,0 +1,47 @@
+name: Test Contributor Lifecycle Bot
+
+on:
+  push:
+    paths:
+      - '.github/scripts/bot-contributor-lifecycle.js'
+      - '.github/scripts/__tests__/bot-contributor-lifecycle.test.js'
+      - '.github/scripts/jest.config.js'
+      - '.github/scripts/package.json'
+      - '.github/workflows/test-bot-contributor-lifecycle.yml'
+  pull_request:
+    paths:
+      - '.github/scripts/bot-contributor-lifecycle.js'
+      - '.github/scripts/__tests__/bot-contributor-lifecycle.test.js'
+      - '.github/scripts/jest.config.js'
+      - '.github/scripts/package.json'
+      - '.github/workflows/test-bot-contributor-lifecycle.yml'
+
+permissions:
+  contents: read
+
+jobs:
+  test:
+    # Pure unit test (no secrets): run on the shared runner, fall back to a hosted
+    # runner for fork PRs that can't use the self-hosted pool.
+    runs-on: ${{ (github.event_name == 'pull_request' && github.event.pull_request.head.repo.fork) && 'ubuntu-latest' || 'hl-sdk-py-lin-md' }}
+    steps:
+      - name: Harden the runner
+        uses: step-security/harden-runner@9af89fc71515a100421586dfdb3dc9c984fbf411 # v2.19.4
+        with:
+          egress-policy: audit
+
+      - name: Checkout repository
+        uses: actions/checkout@df4cb1c069e1874edd31b4311f1884172cec0e10 # v6.0.3
+
+      - name: Setup Node.js
+        uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
+        with:
+          node-version: '20'
+
+      - name: Install script dependencies
+        working-directory: .github/scripts
+        run: npm ci
+
+      - name: Run tests
+        working-directory: .github/scripts
+        run: npx jest __tests__/bot-contributor-lifecycle.test.js --verbose
